Gave WireGraph to its own file.
